Extract unselected things to component
