+++ src/components/info/UnselectedThings/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "UnselectedThings",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./UnselectedThings.js"
+}
